+++ lib/rucio/core/transfer.py
@@ -13,26 +13,23 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import copy
 import datetime
 import logging
 import re
 import time
 import traceback
-from rucio.common.utils import PriorityQueue
 from typing import TYPE_CHECKING
 
 from dogpile.cache import make_region
 from dogpile.cache.api import NoValue
 from sqlalchemy import select, update
 from sqlalchemy.exc import IntegrityError
-from sqlalchemy.sql.expression import false
 
 from rucio.common import constants
 from rucio.common.cache import make_region_memcached
-from rucio.common.config import config_get, config_get_int
+from rucio.common.config import config_get
 from rucio.common.constants import SUPPORTED_PROTOCOLS
-from rucio.common.exception import (InvalidRSEExpression, NoDistance,
+from rucio.common.exception import (InvalidRSEExpression,
                                     RequestNotFound, RSEProtocolNotSupported,
                                     RucioException, UnsupportedOperation)
 from rucio.common.utils import construct_surl
@@ -41,8 +38,9 @@ from rucio.core.account import list_accounts
 from rucio.core.config import get as core_config_get
 from rucio.core.monitor import record_counter
 from rucio.core.request import set_request_state, RequestWithSources, RequestSource
-from rucio.core.rse import list_rses, RseData
+from rucio.core.rse import list_rses, RseCollection
 from rucio.core.rse_expression_parser import parse_expression
+from rucio.core.topology import Topology
 from rucio.db.sqla import models
 from rucio.db.sqla.constants import DIDType, RequestState, RequestType
 from rucio.db.sqla.session import read_session, transactional_session
@@ -50,7 +48,7 @@ from rucio.rse import rsemanager as rsemgr
 from rucio.transfertool.fts3 import FTS3Transfertool
 
 if TYPE_CHECKING:
-    from typing import Any, Callable, Dict, Generator, Iterable, List, Optional, Set, Union
+    from typing import Callable, Dict, Generator, Iterable, List, Optional, Set
     from sqlalchemy.orm import Session
     from rucio.common.types import InternalAccount
 
@@ -75,33 +73,6 @@ class TransferDestination:
 
     def __str__(self):
         return "dst_rse={}".format(self.rse)
-
-
-class _RseLoaderContext:
-    """
-    Helper private class used to dynamically load and cache the rse information
-    """
-    def __init__(self, session):
-        self.session = session
-        self.rse_id_to_data_map = {}
-
-    def rse_data(self, rse_id):
-        rse_data = self.rse_id_to_data_map.get(rse_id)
-        if rse_data is None:
-            rse_data = RseData(rse_id)
-            rse_data.ensure_loaded(load_name=True, load_info=True, load_attributes=True, session=self.session)
-            self.rse_id_to_data_map[rse_id] = rse_data
-        return rse_data
-
-    def ensure_fully_loaded(self, rse_data):
-        if rse_data.name is None or rse_data.info is None or rse_data.attributes is None:
-            cached_rse_data = self.rse_data(rse_data.id)
-            if rse_data.name is None:
-                rse_data.name = cached_rse_data.name
-            if rse_data.info is None:
-                rse_data.info = cached_rse_data.info
-            if rse_data.attributes is None:
-                rse_data.attributes = cached_rse_data.attributes
 
 
 class ProtocolFactory:
@@ -577,153 +548,9 @@ def touch_transfer(external_host, transfer_id, session=None):
         raise RucioException(error.args)
 
 
-@transactional_session
-def get_hops(source_rse_id, dest_rse_id, multihop_rses=None, limit_dest_schemes=None, session=None):
-    """
-    Get a list of hops needed to transfer date from source_rse_id to dest_rse_id.
-    Ideally, the list will only include one item (dest_rse_id) since no hops are needed.
-    :param source_rse_id:       Source RSE id of the transfer.
-    :param dest_rse_id:         Dest RSE id of the transfer.
-    :param multihop_rses:       List of RSE ids that can be used for multihop. If empty, multihop is disabled.
-    :param limit_dest_schemes:  List of destination schemes the matching scheme algorithm should be limited to for a single hop.
-    :returns:                   List of hops in the format [{'source_rse_id': source_rse_id, 'source_scheme': 'srm', 'source_scheme_priority': N, 'dest_rse_id': dest_rse_id, 'dest_scheme': 'srm', 'dest_scheme_priority': N}]
-    :raises:                    NoDistance
-    """
-    if not limit_dest_schemes:
-        limit_dest_schemes = []
-
-    if not multihop_rses:
-        multihop_rses = []
-
-    ctx = _RseLoaderContext(session)
-    shortest_paths = __search_shortest_paths(ctx=ctx, source_rse_ids=[source_rse_id], dest_rse_id=dest_rse_id,
-                                             operation_src='third_party_copy_read', operation_dest='third_party_copy_write',
-                                             domain='wan', multihop_rses=multihop_rses,
-                                             limit_dest_schemes=limit_dest_schemes, session=session)
-
-    result = REGION_SHORT.get('get_hops_dist_%s_%s_%s' % (str(source_rse_id), str(dest_rse_id), ''.join(sorted(limit_dest_schemes))))
-    if not isinstance(result, NoValue):
-        return result
-
-    path = shortest_paths.get(source_rse_id)
-    if path is None:
-        raise NoDistance()
-
-    if not path:
-        raise RSEProtocolNotSupported()
-
-    REGION_SHORT.set('get_hops_dist_%s_%s_%s' % (str(source_rse_id), str(dest_rse_id), ''.join(sorted(limit_dest_schemes))), path)
-    return path
-
-
-def __search_shortest_paths(
-        ctx: _RseLoaderContext,
-        source_rse_ids: "List[str]",
-        dest_rse_id: "List[str]",
-        operation_src: str,
-        operation_dest: str,
-        domain: str,
-        multihop_rses: "List[str]",
-        limit_dest_schemes: "Union[str, List[str]]",
-        inbound_links_by_node: "Optional[Dict[str, Dict[str, str]]]" = None,
-        session: "Optional[Session]" = None,
-) -> "Dict[str, List[Dict[str, Any]]]":
-    """
-    Find the shortest paths from multiple sources towards dest_rse_id.
-    Does a Backwards Dijkstra's algorithm: start from destination and follow inbound links towards the sources.
-    If multihop is disabled, stop after analysing direct connections to dest_rse. Otherwise, stops when all
-    sources where found or the graph was traversed in integrality.
-
-    The inbound links retrieved from the database can be accumulated into the inbound_links_by_node, passed
-    from the calling context. To be able to reuse them.
-    """
-    HOP_PENALTY = config_get_int('transfers', 'hop_penalty', default=10, session=session)  # Penalty to be applied to each further hop
-
-    if multihop_rses:
-        # Filter out island source RSEs
-        sources_to_find = {rse_id for rse_id in source_rse_ids if __load_outgoing_distances_node(rse_id=rse_id, session=session)}
-    else:
-        sources_to_find = set(source_rse_ids)
-
-    next_hop = {dest_rse_id: {'cumulated_distance': 0}}
-    priority_q = PriorityQueue()
-
-    remaining_sources = copy.copy(sources_to_find)
-    priority_q[dest_rse_id] = 0
-    while priority_q:
-        current_node = priority_q.pop()
-
-        if current_node in remaining_sources:
-            remaining_sources.remove(current_node)
-        if not remaining_sources:
-            # We found the shortest paths to all desired sources
-            break
-
-        current_distance = next_hop[current_node]['cumulated_distance']
-        inbound_links = __load_inbound_distances_node(rse_id=current_node, session=session)
-        if inbound_links_by_node is not None:
-            inbound_links_by_node[current_node] = inbound_links
-        for adjacent_node, link_distance in sorted(inbound_links.items(),
-                                                   key=lambda item: 0 if item[0] in sources_to_find else 1):
-            if link_distance is None:
-                continue
-
-            if adjacent_node not in remaining_sources and adjacent_node not in multihop_rses:
-                continue
-
-            try:
-                hop_penalty = int(ctx.rse_data(adjacent_node).attributes.get('hop_penalty', HOP_PENALTY))
-            except ValueError:
-                hop_penalty = HOP_PENALTY
-            new_adjacent_distance = current_distance + link_distance + hop_penalty
-            if next_hop.get(adjacent_node, {}).get('cumulated_distance', 9999) <= new_adjacent_distance:
-                continue
-
-            try:
-                matching_scheme = rsemgr.find_matching_scheme(
-                    rse_settings_src=ctx.rse_data(adjacent_node).info,
-                    rse_settings_dest=ctx.rse_data(current_node).info,
-                    operation_src=operation_src,
-                    operation_dest=operation_dest,
-                    domain=domain,
-                    scheme=limit_dest_schemes if adjacent_node == dest_rse_id and limit_dest_schemes else None
-                )
-                next_hop[adjacent_node] = {
-                    'source_rse_id': adjacent_node,
-                    'dest_rse_id': current_node,
-                    'source_scheme': matching_scheme[1],
-                    'dest_scheme': matching_scheme[0],
-                    'source_scheme_priority': matching_scheme[3],
-                    'dest_scheme_priority': matching_scheme[2],
-                    'hop_distance': link_distance,
-                    'cumulated_distance': new_adjacent_distance,
-                }
-                priority_q[adjacent_node] = new_adjacent_distance
-            except RSEProtocolNotSupported:
-                if next_hop.get(adjacent_node) is None:
-                    next_hop[adjacent_node] = {}
-
-        if not multihop_rses:
-            # Stop after the first iteration, which finds direct connections to destination
-            break
-
-    paths = {}
-    for rse_id in source_rse_ids:
-        hop = next_hop.get(rse_id)
-        if hop is None:
-            continue
-
-        path = []
-        while hop.get('dest_rse_id'):
-            path.append(hop)
-            hop = next_hop.get(hop['dest_rse_id'])
-        paths[rse_id] = path
-    return paths
-
-
 @read_session
 def __create_transfer_definitions(
-        ctx: _RseLoaderContext,
+        topology: Topology,
         protocol_factory: ProtocolFactory,
         rws: RequestWithSources,
         sources: "List[RequestSource]",
@@ -739,10 +566,10 @@ def __create_transfer_definitions(
     Create the transfer definitions for each point-to-point transfer (multi-source, when possible)
     """
     inbound_links_by_node = {}
-    shortest_paths = __search_shortest_paths(ctx=ctx, source_rse_ids=[s.rse.id for s in sources], dest_rse_id=rws.dest_rse.id,
-                                             operation_src=operation_src, operation_dest=operation_dest, domain=domain,
-                                             multihop_rses=multihop_rses, limit_dest_schemes=limit_dest_schemes,
-                                             inbound_links_by_node=inbound_links_by_node, session=session)
+    shortest_paths = topology.search_shortest_paths(source_rse_ids=[s.rse.id for s in sources], dest_rse_id=rws.dest_rse.id,
+                                                    operation_src=operation_src, operation_dest=operation_dest, domain=domain,
+                                                    multihop_rses=multihop_rses, limit_dest_schemes=limit_dest_schemes,
+                                                    inbound_links_by_node=inbound_links_by_node, session=session)
 
     transfers_by_source = {}
     sources_by_rse_id = {s.rse.id: s for s in sources}
@@ -750,8 +577,8 @@ def __create_transfer_definitions(
     for source, list_hops in paths_by_source.items():
         transfer_path = []
         for hop in list_hops:
-            hop_src_rse = ctx.rse_data(hop['source_rse_id'])
-            hop_dst_rse = ctx.rse_data(hop['dest_rse_id'])
+            hop_src_rse = topology.rse_collection[hop['source_rse_id']]
+            hop_dst_rse = topology.rse_collection[hop['dest_rse_id']]
             src = RequestSource(
                 rse_data=hop_src_rse,
                 file_path=source.file_path if hop_src_rse == source.rse else None,
@@ -1077,7 +904,8 @@ def __build_transfer_paths(
 
     Each path is a list of hops. Each hop is a transfer definition.
     """
-    ctx = _RseLoaderContext(session)
+    rse_collection = RseCollection()
+    topology = Topology(rse_collection=rse_collection)
     protocol_factory = ProtocolFactory()
     unavailable_read_rse_ids = __get_unavailable_rse_ids(operation='read', session=session)
     unavailable_write_rse_ids = __get_unavailable_rse_ids(operation='write', session=session)
@@ -1095,9 +923,11 @@ def __build_transfer_paths(
     reqs_unsupported_transfertool = set()
     for rws in requests_with_sources:
 
-        ctx.ensure_fully_loaded(rws.dest_rse)
+        rws.dest_rse = rse_collection.setdefault(rws.dest_rse.id, rws.dest_rse)
+        rws.dest_rse.ensure_loaded(load_name=True, load_info=True, load_attributes=True, session=session)
         for source in rws.sources:
-            ctx.ensure_fully_loaded(source.rse)
+            source.rse = rse_collection.setdefault(source.rse.id, source.rse)
+            source.rse.ensure_loaded(load_name=True, load_info=True, load_attributes=True, session=session)
 
         transfer_schemes = schemes
         if rws.previous_attempt_id and failover_schemes:
@@ -1173,7 +1003,6 @@ def __build_transfer_paths(
             filtered_rses_log += ','.join(filtered_rses[:num_sources_in_logs])
             if len(filtered_rses) > num_sources_in_logs:
                 filtered_rses_log += '... and %d others' % (len(filtered_rses) - num_sources_in_logs)
-        any_source_had_scheme_mismatch = False
         candidate_paths = []
 
         if rws.request_type == RequestType.STAGEIN:
@@ -1184,7 +1013,7 @@ def __build_transfer_paths(
                                                  operation_dest='write',
                                                  protocol_factory=protocol_factory)
         else:
-            paths = __create_transfer_definitions(ctx,
+            paths = __create_transfer_definitions(topology=topology,
                                                   rws=rws,
                                                   sources=filtered_sources,
                                                   multihop_rses=multihop_rses,
@@ -1196,6 +1025,7 @@ def __build_transfer_paths(
                                                   session=session)
 
         sources_without_path = []
+        any_source_had_scheme_mismatch = False
         for source in filtered_sources:
             transfer_path = paths.get(source.rse.id)
             if transfer_path is None:
@@ -1301,68 +1131,6 @@ def __add_compatible_schemes(schemes, allowed_schemes):
                 else:
                     return_schemes.append(scheme_map_scheme)
     return list(set(return_schemes))
-
-
-@transactional_session
-def __load_inbound_distances_node(rse_id, session=None):
-    """
-    Loads the inbound edges of the distance graph for one node.
-    :param rse_id:    RSE id to load the edges for.
-    :param session:   The DB Session to use.
-    :returns:         Dictionary based graph object.
-    """
-
-    result = REGION_SHORT.get('inbound_edges_%s' % str(rse_id))
-    if isinstance(result, NoValue):
-        inbound_edges = {}
-        stmt = select(
-            models.Distance
-        ).join(
-            models.RSE,
-            models.RSE.id == models.Distance.src_rse_id
-        ).where(
-            models.Distance.dest_rse_id == rse_id,
-            models.RSE.deleted == false()
-        )
-        for distance in session.execute(stmt).scalars():
-            if distance.ranking is None:
-                continue
-            ranking = distance.ranking if distance.ranking >= 0 else 0
-            inbound_edges[distance.src_rse_id] = ranking
-        REGION_SHORT.set('inbound_edges_%s' % str(rse_id), inbound_edges)
-        result = inbound_edges
-    return result
-
-
-@transactional_session
-def __load_outgoing_distances_node(rse_id, session=None):
-    """
-    Loads the outgoing edges of the distance graph for one node.
-    :param rse_id:    RSE id to load the edges for.
-    :param session:   The DB Session to use.
-    :returns:         Dictionary based graph object.
-    """
-
-    result = REGION_SHORT.get('outgoing_edges_%s' % str(rse_id))
-    if isinstance(result, NoValue):
-        outgoing_edges = {}
-        stmt = select(
-            models.Distance
-        ).join(
-            models.RSE,
-            models.RSE.id == models.Distance.dest_rse_id
-        ).where(
-            models.Distance.src_rse_id == rse_id,
-            models.RSE.deleted == false()
-        )
-        for distance in session.execute(stmt).scalars():
-            if distance.ranking is None:
-                continue
-            ranking = distance.ranking if distance.ranking >= 0 else 0
-            outgoing_edges[distance.dest_rse_id] = ranking
-        REGION_SHORT.set('outgoing_edges_%s' % str(rse_id), outgoing_edges)
-        result = outgoing_edges
-    return result
 
 
 @read_session
